requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.6
elasticsearch==6.1.1
Flask==0.11.1
Flask-Login==0.4.1
Flask-PyMongo==0.5.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
passlib==1.7.1
pymongo==3.6.1
PyMySQL==0.7.9
SQLAlchemy==1.2.5
urllib3==1.22
virtualenv==15.1.0
Werkzeug==0.11.11
WTForms==2.1
filetype==1.0.1
pillow==2.2.1
